Add test for leading zeroes in a bignum
RFC 7049 Section 2.4.2 states: "Decoders that understand these tags MUST be
able to decode bignums that have leading zeroes." A liberal interpretation of
this statement means that bignums with leading zeroes can be losslessly coerced
into integers of smaller size. Since we aim to follow the robustness principle,
we will accept bignums with leading zeroes so long as the tail of the bignum
fits into a u128.

https://en.wikipedia.org/wiki/Robustness_principle

Signed-off-by: Nathaniel McCallum <npmccallum@redhat.com>