Simplify Value::Integer handling
The Integer type now only holds exactly what can be represented on the
wire as valid CBOR integer types. This means that we can convert from
Value to CBOR and back losslessly. We also `impl From<TYPE> for Value`
for 128-bit integer types. If the integer value doesn't fit in an
Integer, we convert it to the bigint representation. This matches the
regular CBOR behavior.

Signed-off-by: Nathaniel McCallum <npmccallum@redhat.com>