Use `DeserializeOwned` for `from_reader`
Based on #45 I saw that `from_reader` uses the wrong trait bounds for
`T`. It does not support arbitrary types, but only those, which do not
borrow from the input value. This can be expressed with
`DeserializeOwned`.

The `Read` trait always returns owned data and has no concept of
borrowing from the input. Therefore, any type deserialize with the
function also cannot borrow from the input.

You see the same trait bound on `serde_json::from_reader`.

The changes in `tests/` are simply to get the tests compiling
afterwards.