feat: make half dependency optional
Rust is forging ahead with a new f16 type. There's a lot of demand for this
type in the AI space. So it is likely on track for stabilization.

This is the only reason for the half dependency. So we can make it optional.
Users of ciborium-ll must (for now) choose either the half dependency or build
against nightly. Once the new rust f16 type is stabilized, we can remove the
half dependency altogether.

Signed-off-by: Nathaniel McCallum <nathaniel@mccallum.life>